Добавить сервис exchange и заготовку для сервиса transfer, добавить в заготовку сервиса transfer обработку кейса, при котором у пользователя недостаточно средств для перевода самому себе

+++ accounts/src/main/java/ru/ya/dto/BankAccountDto.java
@@ -18,6 +18,6 @@ public class BankAccountDto {
     int id;
     int userId;
     String accountNumber;
-    int accountValue;
+    double accountValue;
     Currency currency;
 }
